Don't exit if dependency is missing package.json (#1)
* Safely handle when no package.json exists

This happens, for example, when a dependency comes from a git repo. Handle with a try/catch. If there is no package.json, we still try to provide name and path in the packageObj. Also, add a test.

* bump version

* Add print and save, BREAKING CHANGE.

This changes what is exported. This allows for exporting a print and save functionality, currently only CSV is supported.